Fixes for issues found during testing: - Extract password operations from IUserService to IPasswordOperations to fix circular dependency between PasswordService and UserService - Fix GORM saving userId by changing userId to a read only property on UserRecord - Remove @EqualsAndHashCode from domain classes - Remove displayName from base UserRecord as it's unused
