+++ userdetails-cognito/src/main/groovy/au/org/ala/userdetails/CognitoPasswordOperations.groovy
@@ -1,0 +1,79 @@
+package au.org.ala.userdetails
+
+import au.org.ala.auth.PasswordResetFailedException
+import au.org.ala.users.UserRecord
+import com.amazonaws.services.cognitoidp.AWSCognitoIdentityProvider
+import com.amazonaws.services.cognitoidp.model.AdminResetUserPasswordRequest
+import com.amazonaws.services.cognitoidp.model.AdminSetUserPasswordRequest
+import com.amazonaws.services.cognitoidp.model.ConfirmForgotPasswordRequest
+import grails.core.GrailsApplication
+import org.apache.commons.codec.digest.HmacAlgorithms
+import org.apache.commons.codec.digest.HmacUtils
+
+class CognitoPasswordOperations implements IPasswordOperations {
+
+    GrailsApplication grailsApplication
+
+    AWSCognitoIdentityProvider cognitoIdp
+    String poolId
+
+    @Override
+    boolean resetPassword(UserRecord user, String newPassword, boolean isPermanent, String confirmationCode) {
+        if(!user || !newPassword) {
+            return false
+        }
+
+        try {
+            if (confirmationCode == null) {
+                def request = new AdminSetUserPasswordRequest()
+                request.username = user.email
+                request.userPoolId = poolId
+                request.password = newPassword
+                request.permanent = isPermanent
+
+                def response = cognitoIdp.adminSetUserPassword(request)
+                return response.getSdkHttpMetadata().httpStatusCode == 200
+            } else {
+                def request = new ConfirmForgotPasswordRequest().withUsername(user.email)
+                request.password = newPassword
+                request.confirmationCode = confirmationCode
+                request.clientId = grailsApplication.config.getProperty('security.oidc.client-id')
+                request.secretHash = calculateSecretHash(grailsApplication.config.getProperty('security.oidc.client-id'),
+                        grailsApplication.config.getProperty('security.oidc.secret'), user.email)
+                def response = cognitoIdp.confirmForgotPassword(request)
+                return response.getSdkHttpMetadata().httpStatusCode == 200
+            }
+        } catch(Exception e) {
+            return false
+        }
+    }
+
+    @Override
+    void resetAndSendTemporaryPassword(UserRecord user, String emailSubject, String emailTitle, String emailBody, String password) throws PasswordResetFailedException {
+        def request = new AdminResetUserPasswordRequest()
+        request.username = user.email
+        request.userPoolId = poolId
+
+        cognitoIdp.adminResetUserPassword(request)
+    }
+
+    static String calculateSecretHash(String userPoolClientId, String userPoolClientSecret, String userName) {
+        try {
+            byte[] rawHmac = new HmacUtils(HmacAlgorithms.HMAC_SHA_256, userPoolClientSecret).hmac("$userName$userPoolClientId")
+            return Base64.getEncoder().encodeToString(rawHmac)
+        } catch (Exception e) {
+            throw new RuntimeException("Error while calculating ")
+        }
+    }
+
+    @Override
+    String getResetPasswordUrl(UserRecord user) {
+        return null
+    }
+
+    @Override
+    String getPasswordResetView() {
+        return "passwordResetCognito"
+    }
+
+}
