+++ userdetails-cognito/src/main/groovy/au/org/ala/userdetails/CognitoUserService.groovy
@@ -1,7 +1,6 @@
 package au.org.ala.userdetails
 
 import au.org.ala.auth.BulkUserLoadResults
-import au.org.ala.auth.PasswordResetFailedException
 import au.org.ala.users.RoleRecord
 import au.org.ala.users.UserPropertyRecord
 import au.org.ala.users.UserRecord
@@ -18,14 +17,10 @@ import com.amazonaws.services.cognitoidp.model.AdminGetUserRequest
 import com.amazonaws.services.cognitoidp.model.AdminGetUserResult
 import com.amazonaws.services.cognitoidp.model.AdminListGroupsForUserRequest
 import com.amazonaws.services.cognitoidp.model.AdminRemoveUserFromGroupRequest
-import com.amazonaws.services.cognitoidp.model.AdminRemoveUserFromGroupResult
-import com.amazonaws.services.cognitoidp.model.AdminResetUserPasswordRequest
 import com.amazonaws.services.cognitoidp.model.AdminSetUserMFAPreferenceRequest
-import com.amazonaws.services.cognitoidp.model.AdminSetUserPasswordRequest
 import com.amazonaws.services.cognitoidp.model.AdminUpdateUserAttributesRequest
 import com.amazonaws.services.cognitoidp.model.AssociateSoftwareTokenRequest
 import com.amazonaws.services.cognitoidp.model.AttributeType
-import com.amazonaws.services.cognitoidp.model.ConfirmForgotPasswordRequest
 import com.amazonaws.services.cognitoidp.model.CreateGroupRequest
 import com.amazonaws.services.cognitoidp.model.GetGroupRequest
 import com.amazonaws.services.cognitoidp.model.GetUserRequest
@@ -43,8 +38,6 @@ import com.nimbusds.oauth2.sdk.token.AccessToken
 import com.amazonaws.services.cognitoidp.model.VerifySoftwareTokenRequest
 import grails.web.servlet.mvc.GrailsParameterMap
 import groovy.util.logging.Slf4j
-import org.apache.commons.codec.digest.HmacAlgorithms
-import org.apache.commons.codec.digest.HmacUtils
 import org.apache.commons.lang3.NotImplementedException
 import org.springframework.beans.factory.annotation.Value
 
@@ -59,7 +52,6 @@ class CognitoUserService implements IUserService {
 
     EmailService emailService
     TokenService tokenService
-    def grailsApplication
 
     AWSCognitoIdentityProvider cognitoIdp
     String poolId
@@ -214,7 +206,6 @@ class CognitoUserService implements IUserService {
 
         def user = new UserRecord<String>(
                 id: attributes['name'] ?: userType.username,
-                userId: userType.username,
                 dateCreated: userType.userCreateDate, lastUpdated: userType.userLastModifiedDate,
                 activated: userType.userStatus == "CONFIRMED", locked: !userType.enabled,
                 firstName: attributes['given_name'], lastName: attributes['family_name'],
@@ -286,7 +277,6 @@ class CognitoUserService implements IUserService {
                     lastName: attributes.find { it.key == 'family_name' }.value,
                     email: attributes.find { it.key == 'email' }.value,
                     userName: userResponse.user.username,
-                    userId: userResponse.user.username,
                     userProperties: userProperties
             )
 
@@ -304,15 +294,6 @@ class CognitoUserService implements IUserService {
             return user
         }
         return null
-    }
-
-    @Override
-    void resetAndSendTemporaryPassword(UserRecord user, String emailSubject, String emailTitle, String emailBody, String password) throws PasswordResetFailedException {
-        def request = new AdminResetUserPasswordRequest()
-        request.username = user.email
-        request.userPoolId = poolId
-
-        cognitoIdp.adminResetUserPassword(request)
     }
 
     @Override
@@ -355,7 +336,6 @@ class CognitoUserService implements IUserService {
 
             UserRecord user = new UserRecord<String>(
                     id: attributes['name'] ?: userResponse.username,
-                    userId: userResponse.username,
                     dateCreated: userResponse.userCreateDate, lastUpdated: userResponse.userLastModifiedDate,
                     activated: userResponse.userStatus == "CONFIRMED", locked: !userResponse.enabled,
                     firstName: attributes['given_name'], lastName: attributes['family_name'],
@@ -404,7 +384,6 @@ class CognitoUserService implements IUserService {
 
             UserRecord user = new UserRecord<String>(
                     id: attributes['name'] ?: userResponse.username,
-                    userId: userResponse.username,
 //                dateCreated: userResponse.userCreateDate, lastUpdated: userResponse.userLastModifiedDate,
 //                activated: userResponse.userStatus == "CONFIRMED", locked: !userResponse.enabled,
                     firstName: attributes['given_name'], lastName: attributes['family_name'],
@@ -419,11 +398,6 @@ class CognitoUserService implements IUserService {
             log.error(e.getMessage())
             return null
         }
-    }
-
-    @Override
-    String getResetPasswordUrl(UserRecord user) {
-        return null
     }
 
     @Override
@@ -641,42 +615,6 @@ class CognitoUserService implements IUserService {
     }
 
     @Override
-    boolean resetPassword(UserRecord user, String newPassword, boolean isPermanent, String confirmationCode) {
-        if(!user || !newPassword) {
-            return false
-        }
-
-        try {
-            if (confirmationCode == null) {
-                def request = new AdminSetUserPasswordRequest()
-                request.username = user.email
-                request.userPoolId = poolId
-                request.password = newPassword
-                request.permanent = isPermanent
-
-                def response = cognitoIdp.adminSetUserPassword(request)
-                return response.getSdkHttpMetadata().httpStatusCode == 200
-            } else {
-                def request = new ConfirmForgotPasswordRequest().withUsername(user.email)
-                request.password = newPassword
-                request.confirmationCode = confirmationCode
-                request.clientId = grailsApplication.config.getProperty('security.oidc.client-id')
-                request.secretHash = calculateSecretHash(grailsApplication.config.getProperty('security.oidc.client-id'),
-                        grailsApplication.config.getProperty('security.oidc.secret'), user.email)
-                def response = cognitoIdp.confirmForgotPassword(request)
-                return response.getSdkHttpMetadata().httpStatusCode == 200
-            }
-        }catch(Exception e){
-            return false
-        }
-    }
-
-    @Override
-    String getPasswordResetView() {
-        return "passwordResetCognito"
-    }
-
-    @Override
     def sendAccountActivation(UserRecord user) {
         //this email can be sent via cognito
         //emailService.sendCognitoAccountActivation(user)
@@ -721,15 +659,6 @@ class CognitoUserService implements IUserService {
         def response = cognitoIdp.adminSetUserMFAPreference(mfaRequest)
         if (response.sdkHttpMetadata.httpStatusCode != 200) {
             throw new RuntimeException("Couldn't set MFA preference")
-        }
-    }
-
-    static String calculateSecretHash(String userPoolClientId, String userPoolClientSecret, String userName) {
-        try {
-            byte[] rawHmac = new HmacUtils(HmacAlgorithms.HMAC_SHA_256, userPoolClientSecret).hmac("$userName$userPoolClientId")
-            return Base64.getEncoder().encodeToString(rawHmac)
-        } catch (Exception e) {
-            throw new RuntimeException("Error while calculating ")
         }
     }
