+++ userdetails-gorm/grails-app/domain/au/org/ala/userdetails/gorm/User.groovy
@@ -20,15 +20,13 @@ import groovy.transform.EqualsAndHashCode
 
 import java.sql.Timestamp
 
-@EqualsAndHashCode(includes = 'id')
+@EqualsAndHashCode
 class User extends UserRecord<Long> implements Serializable {
 
     static hasMany =  [
             userRoles: UserRole,
             userProperties: UserProperty
     ]
-
-    String userId
 
     String firstName
     String lastName
@@ -48,14 +46,13 @@ class User extends UserRecord<Long> implements Serializable {
 
     String displayName
 
-    Collection<UserRole> userRoles
-    Collection<UserProperty> userProperties
+//    Collection<UserRole> userRoles
+//    Collection<UserProperty> userProperties
 
     static mapping = {
         table 'users'
 
         id (generator:'identity', column:'userid', type:'long')
-        userId column:'userid', updatable: false, insertable: false, type: 'string'
 
         userName column:  'username'
         firstName column:  'firstname'
@@ -76,6 +73,11 @@ class User extends UserRecord<Long> implements Serializable {
         lastLogin nullable: true
         tempAuthKey nullable: true
         displayName nullable: true
+    }
+
+    @Override
+    String getUserId() {
+        return this.id?.toString()
     }
 
     static List<String[]> findNameAndEmailWhereEmailIsNotNull() {
