+++ userdetails-plugin/src/test/groovy/au/org/ala/userdetails/RegistrationControllerSpec.groovy
@@ -194,7 +194,7 @@ class RegistrationControllerSpec extends UserDetailsSpec implements ControllerUn
         controller.updatePassword()
 
         then:
-        1 * userService.resetPassword(user, password, _, _)
+        1 * passwordService.resetPassword(user, password, _, _)
         1 * passwordService.validatePassword(user.email, password) >> new RuleResult(true)
         1 * passwordService.resetPassword(user, password)
         1 * userService.clearTempAuthKey(user)
@@ -238,7 +238,7 @@ class RegistrationControllerSpec extends UserDetailsSpec implements ControllerUn
         1 * userService.isEmailRegistered(email) >> false
         1 * passwordService.validatePassword(email, password) >> new RuleResult(true)
         1 * userService.registerUser(_) >> { def user = new UserRecord(params); user.tempAuthKey = authKey; user }
-        1 * userService.resetPassword(_, password, _, _)
+        1 * passwordService.resetPassword(_, password, _, _)
         1 * emailService.sendAccountActivation(_, authKey)
         0 * _ // no other interactions
         response.redirectedUrl == '/registration/accountCreated'
@@ -312,7 +312,7 @@ class RegistrationControllerSpec extends UserDetailsSpec implements ControllerUn
         then:
         1 * recaptchaClient.verify(secretKey, null, '127.0.0.1') >> { Calls.response(new RecaptchaResponse(false, null, null, ['missing-input-response'])) }
         0 * userService.registerUser(_)
-        0 * userService.resetPassword(_, _, _, _)
+        0 * passwordService.resetPassword(_, _, _, _)
         0 * emailService.sendAccountActivation(_, _)
         0 * _ // no other interactions
         view == '/registration/createAccount'
